+++ Core/Core/Classes/Responces/CoinResponse.swift
@@ -24,7 +24,7 @@ public struct CoinResponse {
     public let currentPrice: Double
     public let marketCap: Int
     public let marketCapRank: Int
-    public let totalVolume: Int
+    public let totalVolume: Double
     public let previousDayHighestPrice: Double
     public let previousDayLowestPrice: Double
     public let priceChange24h: Double
